# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Survivor League API
  description: >
    REST API for the College Football Survivor Pool app. This spec is the
    source of truth for the shared TypeScript client generated into
    packages/api-client (consumed by both web and mobile).

    Phase 1 added auth (register/login/refresh/logout) and the current-user
    endpoints. Phase 2 added league CRUD, membership, the invite-code join
    flow, and the conferences list. Phase 3 added read-only schedule
    endpoints (weeks/games/teams, populated by a CFBD sync) and the first
    site-admin endpoints (triggering/viewing that sync). Phase 4 added pick
    submission/retrieval with server-enforced per-game locking and the
    no-repeat-team rule, plus pre-lock pick-hiding for other league members.
    Phase 5 adds the leaderboard endpoint — the grading/elimination pipeline
    and the live poll loop that drives it are internal (internal/grading,
    internal/livepoll), not client-facing API surface. Phase 6 adds the
    commissioner buy-back endpoint (one-time-ever reinstatement of an
    eliminated member). Phase 7 adds notifications: device-token
    registration and notification preferences (the notification_outbox
    dispatcher, Expo Push/Resend delivery, and the five trigger types —
    pick_reminder, eliminated, survived, mass_wipeout, buyback — are
    internal, internal/notify, not client-facing API surface beyond these
...
    (GET /admin/leagues, GET /admin/users), user disable/enable, a
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /leagues:
    post:
      operationId: createLeague
      summary: Create a new league
      description: >
        Creates a league locked to the given conference/season_year for
        its lifetime (both immutable after creation) with a fresh unique
        invite_code, and — in the same transaction, with no separate join
        step — the requester's own league_memberships row as
        role=commissioner, is_contestant=true, status=active.
      tags:
        - Leagues
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeagueRequest"
      responses:
        "201":
          description: League created; requester is its commissioner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/League"
        "400":
          description: Missing name, season_year out of range, or conference not recognized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, invalid, or expired access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listMyLeagues
      summary: List leagues the requester belongs to
      description: >
        Leagues the requester has a non-removed membership in, each with
        their role/is_contestant/status in that league.
      tags:
        - Leagues
      responses:
        "200":
          description: The requester's leagues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/League"
        "401":
          description: Missing, invalid, or expired access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /leagues/{id}/invite/regenerate:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: regenerateInviteCode
      summary: Regenerate a league's invite code (commissioner only)
      description: >
        Requires role=commissioner (requireCommissioner). Generates a
        fresh unique code and overwrites the stored one — the old code
        stops working immediately (it isn't a multi-valued list, just a
        single column).
      tags:
        - Invites
      responses:
        "200":
          description: The league's new invite code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteCodeResponse"
        "403":
          description: Not the league's commissioner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: League not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    CreateLeagueRequest:
      type: object
      required:
        - name
        - season_year
        - conference
      properties:
        name:
          type: string
        season_year:
          type: integer
          minimum: 2000
          maximum: 2100
        conference:
          type: string
          description: Must exactly (case-sensitive) match one of the names returned by GET /conferences.
...
    MembershipSummary:
      type: object
      description: The requester's own role/status within a league.
      required:
        - id
        - role
        - is_contestant
        - status
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [commissioner, player]
        is_contestant:
          type: boolean
        status:
          type: string
          enum: [active, eliminated]
    League:
      type: object
      description: >
        Returned by every endpoint that hands back a league together with
        the requester's membership in it (create, list, get, update,
        join). Deliberately omits invite_code — that's only exposed via
        GET/POST /leagues/{id}/invite....
      required:
        - id
        - name
        - season_year
        - conference
        - commissioner_user_id
        - status
        - created_at
        - membership
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        season_year:
          type: integer
        conference:
          type: string
        commissioner_user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [active]
        created_at:
          type: string
          format: date-time
        membership:
          $ref: "#/components/schemas/MembershipSummary"
...
    InviteCodeResponse:
      type: object
      required:
        - invite_code
      properties:
        invite_code:
          type: string
...